Enable more warnings.

+++ sysinit.c
@@ -30,6 +30,7 @@
  */
 
 #include "chip.h"
+#include <string.h>
 
 //#define USE_EXTERNAL_OSCILLATOR 1
 
@@ -179,7 +180,7 @@ STATIC void SystemSetupClocking(void)
 /* Sets up system pin muxing */
 STATIC void SystemSetupMuxing(void)
 {
-	int i;
+	size_t i;
 
 	/* Enable IOCON clock */
 	Chip_Clock_EnablePeriphClock(SYSCTL_CLOCK_IOCON);
